lab7 / lect 8

+++ lectures/Lecture8Fix.java
@@ -13,20 +13,21 @@ public class Lecture8Fix {
   public static void main(String[] args) {
    // challenge #1
    short a=4,b=5,c;
-   c = (a + b); 
+   c = (short)(a + b); 
    System.out.println("a + b = c: " + a +"+"+b+"="+c);
    /*
    // challenge #2 
    int aa;
    long bb;
    bb = 12; 
-   aa = bb + 15;
+   aa = (int)bb + 15;
    System.out.println("15 + b = a: " + "15+"+bb+"="+aa);
+   
    // challenge #3 
    int sumGrades = 789;
    int numStudents = 10;
    double classAvg;
-   classAvg = sumGrades/numStudents;
+   classAvg = (double)sumGrades/numStudents;
    System.out.println("class average " + classAvg);
    */
   }
